Fix autogenerated append to original _headers file
Vercel hard links the public directory so when next-on-pages appends to
the vercel output directory, it is also appending to the original
`public/_headers` file too.

This commit fixes the problem by copying the contents to memory before
unlinking the file and creating a new (non-linked) file.

I was able to test out this patch and saw it worked for when `_headers`
exists and when it is absent.

Closes #848