Add CLI option to ignore invalid functions
The presence of invalid functions shouldn't force a build failure as a
deployment can still be valid (and even correct as outlined in #845)
with invalid functions.

This commit adds a `--force` cli option (bike-shedding welcomed) that
allows users to opt to ignore invalid functions (and skip more checks in
the future?).

Closes #845